requirements.txt
africastalking
customtkinter
Flask
numpy
opencv_python
Pillow
playsound
ultralytics
# gunicorn
# mediapipe
